requirements.txt
pdfplumber
python-telegram-bot
redis
python-dotenv